+++ Cargo.toml
@@ -25,6 +25,7 @@ anyhow = "1.0.98"
 log = "0.4.27"
 mdbook-core = { path = "crates/mdbook-core" }
 mdbook-preprocessor = { path = "crates/mdbook-preprocessor" }
+mdbook-renderer = { path = "crates/mdbook-renderer" }
 mdbook-summary = { path = "crates/mdbook-summary" }
 memchr = "2.7.5"
 pulldown-cmark = { version = "0.10.3", default-features = false, features = ["html"] } # Do not update, part of the public api.
@@ -63,6 +64,7 @@ hex = "0.4.3"
 log.workspace = true
 mdbook-core.workspace = true
 mdbook-preprocessor.workspace = true
+mdbook-renderer.workspace = true
 mdbook-summary.workspace = true
 memchr.workspace = true
 opener = "0.8.1"
